move password view model to settings folder
