+++ Src/MoneyManager.Windows/Controls/PasswordUserControl.xaml.cs
@@ -1,5 +1,5 @@
 ﻿using Cirrious.CrossCore;
-using MoneyManager.Core.ViewModels;
+using MoneyManager.Core.ViewModels.SettingViews;
 
 namespace MoneyManager.Windows.Controls
 {
